Framework: move RichText component to rich-text package
