+++ packages/rich-text/README.md
@@ -273,6 +273,10 @@ _Returns_
 
 -   `Object`: A new value with replacements applied.
 
+<a name="RichText" href="#RichText">#</a> **RichText**
+
+Undocumented declaration.
+
 <a name="slice" href="#slice">#</a> **slice**
 
 Slice a Rich Text value from `startIndex` to `endIndex`. Indices are
