+++ packages/rich-text/src/component/aria.js
