+++ packages/rich-text/src/component/test/index.js
